requirements.txt
openai>=1.50.0
streamlit>=1.37.0
python-dotenv>=1.0.1
pytest>=8.2.0
